Fix crash

+++ src/physics/constraints/type_batch_alloc.rs
@@ -575,7 +575,7 @@ pub unsafe fn allocate_in_type_batch_for_fallback(
             let mut next_probe_index =
                 ((HashHelper::rehash(handle.0) & 0x7FFF_FFFF) as usize) % last_bundle_index;
             let bundle_jump = bundle_count / PROBE_LOCATION_COUNT;
-            let remainder = last_bundle_index - bundle_jump * PROBE_LOCATION_COUNT;
+            let remainder = last_bundle_index as isize - (bundle_jump * PROBE_LOCATION_COUNT) as isize;
             for probe_index in 0..PROBE_LOCATION_COUNT {
                 if probe_bundle_for_fallback(
                     body_references_ptr,
@@ -590,7 +590,7 @@ pub unsafe fn allocate_in_type_batch_for_fallback(
                     break;
                 }
                 next_probe_index += bundle_jump;
-                if probe_index < remainder {
+                if (probe_index as isize) < remainder {
                     next_probe_index += 1;
                 }
                 if next_probe_index >= bundle_count {
